% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{apply_habitat_actions}
\alias{apply_habitat_actions}
\title{Apply Habitat Actions to Update Habitat Conditions}
\usage{
apply_habitat_actions(
  scenario,
  params,
  starting_habitat,
  starting_hydrology,
  species
)
}
\arguments{
\item{scenario}{A list containing the actions to be applied to the habitat.}

\item{params}{A list of parameters including prey density, prop_high_predation, contact_points, delta_contact_points, delta_prop_high_predation.}

\item{starting_habitat}{A string specifying the starting habitat type.}

\item{starting_hydrology}{A string specifying the starting hydrology type.}

\item{species}{A string specifying the species for which habitat is being updated.}
}
\value{
A list containing the updated habitat conditions for spawning habitat, inchannel habitat for fry and juvenile stages, floodplain habitat, yolo habitat, sutter habitat, delta habitat, weeks flooded, average temperature, and degree days.
}
\description{
This function updates habitat conditions based on the provided scenario, parameters, starting habitat, starting hydrology, and species.
}
\details{
The function performs several actions based on the provided scenario:
\itemize{
  \item Updates habitat conditions based on the species and starting habitat type actions 1 - 3.
  \item Applies action 4 to update yolo habitat based on rice lands salmon rearing practice.
  \item Applies action 5 to increase prey density.
  \item Applies action 6 to decrease predation by scaling contact points.
}
}
