% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{apply_harvest_actions}
\alias{apply_harvest_actions}
\title{Apply Harvest Actions to Update Harvest Parameters}
\usage{
apply_harvest_actions(scenario, params, species)
}
\arguments{
\item{scenario}{A list containing the actions to be applied to the harvest parameters.}

\item{params}{A list of parameters including restrict_harvest_to_hatchery_ocean, restrict_harvest_to_hatchery_trib, ocean_harvest_percentage, tributary_harvest_percentage, crr_scaling, no_cohort_harvest_years, intelligent_crr_harvest, intelligent_hatchery_harvest, preserve_tribal_harvest.}

\item{species}{A string specifying the species for which harvest parameters are being updated.}
}
\value{
A list containing the updated harvest parameters.
}
\description{
This function updates harvest parameters based on the provided scenario, parameters, and species.
}
\details{
The function performs several actions based on the provided scenario:
\itemize{
  \item Updates harvest parameters based on the baseline values provided in params.
  \item Applies action 11 to preserve tribal harvest.
  \item Applies action 12 to set in-river harvest only.
  \item Applies action 13 to set ocean harvest only.
  \item Applies action 14 to enable intelligent CRR harvest.
  \item Applies action 15 to enable intelligent habitat harvest.
  \item Applies action 16 to restrict ocean harvest to hatchery fish only.
  \item Applies action 17 to restrict tributary harvest to hatchery fish only.
}
}
