% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{decay_rate}
\alias{decay_rate}
\alias{spawn_decay_rate}
\alias{rear_decay_rate}
\title{Decay Rates}
\usage{
spawn_decay_rate

rear_decay_rate
}
\description{
An ordered vector of minimum decay rate for each watershed
}
\keyword{datasets}
