% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-scenario.R
\name{get_action_matrices}
\alias{get_action_matrices}
\title{Get Action Matrices}
\usage{
get_action_matrices(scenario_df)
}
\arguments{
\item{scenario_df}{scenario dataframe}
}
\description{
Converts scenario dataframe into a list of matrices [31 watersheds, years]
   for each action type. The matrix values are the number of units of effort applied
   for that action type within a watershed for that year.
}
\details{
The \code{scenario_df} is a dataframe with each row representing a scenario action.
The dataframe must contain the following columns:
\itemize{
  \item \strong{watershed} - The name or index for a watershed, ex. "Upper Sacramento River" or 1
  \item \strong{action} - The action taken represented by the code 1 - 5, ex: 3 to represent adding inchannel rearing
  \item \strong{start_year} - The simulation year the action begins
  \item \strong{end_year} - The simulation year the action ends
  \item \strong{units_of_effort} - number of action units taken, ex: .5 or 1
}
}
\section{Actions and Units of Effort}{

\itemize{
  \item \strong{1} - Do nothing
  \item \strong{2} - Add 1 acre of spawning habitat
  \item \strong{3} - Add 2 acres of inchannel rearing habitat
  \item \strong{4} - Add 3 acres of floodplain rearing habitat
  \item \strong{5} - Increase rearing survival by 5\%

}
}

