% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_habitat}
\alias{max_habitat}
\alias{max_spawn_area}
\alias{max_rear_area}
\title{Theoretical Maximum Habitat}
\usage{
max_spawn_area

max_rear_area
}
\description{
A list for each species containing an ordered vector of the
maximum possible suitable area for each watershed
}
\details{
Suitable habitat modeling captures degraded existing habitat conditions.
For the purposes of scenario development, we assumed that existing maximum habitat
(which occurs at the most suitable flow conditions in the watershed) can be increased
by a maximum of 100% over the twenty year simulation period. This reflects both
feasibility of habitat restoration over twenty years as well as physical limitations
(e.g. homes, levees, bridges, etc.) on habitat conditions.
}
\examples{
# Subset by run to find max spawn and rear habitat area
DSMscenario::max_spawn_area$FALL
DSMscenario::max_rear_area$FALL
}
\keyword{datasets}
