% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_type_lookup}
\alias{param_type_lookup}
\title{Param Type Lookup}
\format{
A dataframe with 2 columns and 46 rows.

A names list containing scenario action tables.
}
\usage{
param_type_lookup

param_type_lookup
}
\description{
A dataframe with 2 columns and 46 rows. 
2 column describe parameter name, and parameter structure. This dataframe is used to update parameters.

A list that contains all the scenarios currently considered in the R2R process.
}
\keyword{datasets}
