% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulated_watersheds}
\alias{regulated_watersheds}
\title{Regulated Watersheds}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
regulated_watersheds
}
\description{
An ordered vector of boolean values, 1 represents a major dam
  on the watershed.
}
\keyword{datasets}
