% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawning_decay_multiplier}
\alias{spawning_decay_multiplier}
\title{Spawning Decay Multiplier}
\format{
list of 4 31 x 22 x 12 arrays.
}
\usage{
spawning_decay_multiplier
}
\description{
A list of 4 arrays for each run "fr", "sr", "wr", "lfr". Each list element
is an array 31 x 22 x 12 containing spawning decay by watershed, year and month.
}
\details{
this data was re-exported from DSMhabitat::spawning_decay_multiplier
}
\keyword{datasets}
