% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Species}
\format{
An object of class \code{list} of length 5.
}
\usage{
species
}
\description{
A helper list object for setting the species value when running \code{load_scenario}
}
\examples{
habitats <- list(
  spawning_habitat = fallRunDSM::params$spawning_habitat,
  inchannel_habitat_fry = fallRunDSM::params$inchannel_habitat_fry,
  inchannel_habitat_juvenile = fallRunDSM::params$inchannel_habitat_juvenile,
  floodplain_habitat = fallRunDSM::params$floodplain_habitat,
  weeks_flooded = fallRunDSM::params$weeks_flooded
)

scenario_df <- data.frame(watershed = c("Upper Sacramento River",
                                        "Upper Sacramento River",
                                        "American River", "Feather River",
                                        "Lower-mid Sacramento River",
                                        "Battle Creek", "Butte Creek",
                                        "Deer Creek", "Stanislaus River"),
                          action = c(3, 3, 3, 3, 3, 3, 3, 3, 3),
                          start_year = c(1980, 1990, 1980, 1980, 1980, 1990,
                                         1990, 1990, 1990),
                          end_year = c(1989, 1999, 1989, 1989, 1989, 1999,
                                         1999, 1999, 1999),
                          units_of_effort = c(2, 1, 1, 1, 1, 1, 1, 1, 1))

scenario <- load_scenario(scenario = DSMscenario::scenarios$ONE,
                          habitat_inputs = habitats,
                          species = DSMscenario::species$FALL,
                          spawn_decay_rate = DSMscenario::spawn_decay_rate,
                          rear_decay_rate = DSMscenario::rear_decay_rate,
                          stochastic = TRUE)
}
\keyword{datasets}
