% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_groups}
\alias{watershed_groups}
\title{SIT Watershed Groupings}
\format{
An object of class \code{numeric} of length 31.
}
\usage{
watershed_groups
}
\description{
SIT defined watershed diversity groups developed from the Central Valley Chinook Salmon and Steelhead Recovery Plan \href{https://archive.fisheries.noaa.gov/wcr/publications/recovery_planning/salmon_steelhead/domains/california_central_valley/cv_chin_stlhd_r_plan_fs_071614.pdf}{NOAA 2014}
}
\keyword{datasets}
